refactor: reuse exsisting types, keep naming consistent
