+++ src/routes/graphql/types/memberType.ts
@@ -6,19 +6,20 @@ import {
   GraphQLObjectType,
 } from 'graphql';
 import { MemberType } from '@prisma/client';
+import { MemberTypeId } from '../../member-types/schemas.js';
 
-export const MemberTypeId = new GraphQLEnumType({
+export const MemberTypeIdType = new GraphQLEnumType({
   name: 'MemberTypeId',
-  values: {
-    BASIC: { value: 'BASIC' },
-    BUSINESS: { value: 'BUSINESS' },
-  },
+  values: Object.fromEntries(
+    Object.keys(MemberTypeId)
+      .map((key) => [key, { value: MemberTypeId[key] as string }]),
+  ),
 });
 
 export const MemberTypeType = new GraphQLObjectType<MemberType>({
   name: 'MemberType',
   fields: {
-    id: { type: new GraphQLNonNull(MemberTypeId) },
+    id: { type: new GraphQLNonNull(MemberTypeIdType) },
     discount: { type: new GraphQLNonNull(GraphQLFloat) },
     postsLimitPerMonth: { type: new GraphQLNonNull(GraphQLInt) },
   },
